A program for a small trucking company
Record the money we make from the work along with the truck name to an
excel file. Program can analyze the data and generate a report on how
much money we have made. Program also checks for invalid data. Program
makes use of python lists and can take the name of a file and append
.csv to it.